re-enable OverviewPanel and renamed BurnDownChartContainer (also per default hidden)
